Seção 17- Palestrantes e redes sociais frontend

+++ Projetos/ProEventos/Back/src/ProEventos.API/Helpers/Util.cs
@@ -17,9 +17,12 @@ namespace ProEventos.API.Helpers
 
         public void DeleteImage(string imageName, string destino)
         {
-            var imagePath = Path.Combine(_hostEnvironment.ContentRootPath, @$"Resources/{destino}", imageName);
-            if (File.Exists(imagePath))
-                File.Delete(imagePath);
+            if (!string.IsNullOrEmpty(imageName))
+            {
+                var imagePath = Path.Combine(_hostEnvironment.ContentRootPath, @$"Resources/{destino}", imageName);
+                if (File.Exists(imagePath))
+                    File.Delete(imagePath);
+            }
         }
 
         public async Task<string> SaveImage(IFormFile imageFile, string destino)
